# mktname9/marketplace-nine/docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"marketplace-nine","description":"Marketplace Nine","contact":{"name":"maulopes","email":"mauricio.lopes@mercadolivre.com"},"version":"0.0.1"},"servers":[{"url":"http://localhost:8080","description":"Scope local"}],"paths":{"/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProductTitle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/products":{"get":{"tags":["product-controller"],"operationId":"findAll","parameters":[{"name":"title","in":"query","required":false,"schema":{"type":"string"}},{"name":"minPrice","in":"query","required":false,"schema":{"type":"number"}},{"name":"maxPrice","in":"query","required":false,"schema":{"type":"number"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}},"post":{"tags":["product-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/products/category":{"get":{"tags":["category-controller"],"operationId":"getAllCategoryNames","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/products/category/{category}":{"get":{"tags":["product-controller"],"operationId":"getProductsByCategory","parameters":[{"name":"category","in":"path","required":true,"schema":{"$ref":"#/components/schemas/Category"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Category":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"Product":{"required":["description","image","price","title"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"price":{"type":"number"},"description":{"type":"string"},"image":{"type":"string"},"category":{"$ref":"#/components/schemas/Category"}}}}}}